Move hoc into common folder
