+++ plugins/common/hoc/__tests__/with-details.test.js
