+++ plugins/common/hoc/with-details.js
