+++ plugins/common/hoc/with-store.js
