+++ plugins/events/src/modules/blocks/classic-event-details/organizer/item.js
@@ -12,7 +12,7 @@ import { isEmpty } from 'lodash';
 /**
  * Import internal dependencies
  */
-import { withDetails } from '@@plugins/events/editor/hoc';
+import { withDetails } from '@@plugins/common/hoc';
 import { Loading } from 'elements';
 import { Actions, Name } from './index';
