+++ plugins/events/src/modules/blocks/event-organizer/index.js
@@ -13,7 +13,7 @@ import { __ } from '@wordpress/i18n';
  */
 import Organizer from './block';
 import { Icons } from 'elements';
-import { withStore } from '@@plugins/events/editor/hoc';
+import { withStore } from '@@plugins/common/hoc';
 import { ORGANIZER } from '@@plugins/events/editor/post-types';
 
 export default {
